Multi-user access on the client and logout implementation on the server

+++ game/2D MMORPG/SIM SERVER/Project1/objectManager.cpp
@@ -3,9 +3,9 @@
 
 bool ObjectManager::calcInRange(Object& objA, Object& objB)
 {
-    int objA_x = objA.m_x.load(), objA_y = objA.m_y.load();
-    int objB_x = objB.m_x.load(), objB_y = objB.m_y.load();
-    int distance = (objA_x - objB_x) * (objA_x - objB_x) +
+    double objA_x = objA.m_x.load(), objA_y = objA.m_y.load();
+    double objB_x = objB.m_x.load(), objB_y = objB.m_y.load();
+    double distance = (objA_x - objB_x) * (objA_x - objB_x) +
         (objA_y - objB_y) * (objA_y - objB_y);
     if (distance <= VIEWRANGE * VIEWRANGE) return true;
     else return false;
